requirements.txt
langchain>=0.3.0
langchain-core>=1.2.0
langchain-openai>=1.1.0
langchain-google-genai>=2.0.0
langgraph>=0.3.0
streamlit>=1.40.0
python-dotenv>=1.0.0
pytest>=8.0.0
